- Added context menu message; - Moved InvalideWebAppTreeItem to it's own file; - Added desciption property; - Cleaned up unused vars;
